refactor: rename <PricingPageLayout> to <PageLayout>
