+++ pages/[...slug].tsx
@@ -2,7 +2,7 @@ import { MDXRemote } from "next-mdx-remote"
 
 import MDXComponents from "@/components/MDXComponents"
 import PageTitle from "@/components/PageTitle"
-import PricingPageLayout from "@/layouts/PricingPageLayout"
+import PricingPageLayout from "@/layouts/PageLayout"
 import { getFileBySlug } from "@/lib/mdx"
 import getFilesRecursively from "@/lib/utils/files"
